Move models to their own file, create db
